pyproject.toml
[project]
name = "vector-search-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.8.0",
    "google-genai>=1.20.0",
    "pymilvus[model]>=2.5.11",
    "python-dotenv>=1.1.0",
]
